refactor(#292): split worktree-orchestrator.ts into helper modules
